Potential fix for code scanning alert no. 11

+++ arch/arm/mach-sunxi/dram_sun8i_a83t.c
@@ -468,6 +468,6 @@ unsigned long sunxi_dram_init(void)
 	else
 		writel(0x00000201, &mctl_ctl->odtmap);
 
-	return para.page_size * (para.bus_width / 8) *
-		(1 << (para.bank + para.rank + para.rows));
+	return ((unsigned long)para.page_size) * (para.bus_width / 8) *
+		(1UL << (para.bank + para.rank + para.rows));
 }
